test: refactor
